seniorsync/backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
openai==2.8.1
python-dotenv==1.2.1
twilio==9.8.7
requests==2.32.5
firebase-admin==7.1.0
python-multipart==0.0.17
